ml_model: make templates private
